+++ src/ui/src/components/filters/general-model-property-selector.vue
@@ -51,7 +51,7 @@
     },
     methods: {
       async confirm() {
-        this.handler(this.selected)
+        this.handler(this.$refs?.addConditionComp?.localSelected ?? this.selected)
         this.close()
       }
     }
