+++ src/ui/src/components/filters/property-selector.vue
@@ -11,81 +11,28 @@
 -->
 
 <template>
-  <bk-dialog
-    v-model="isShow"
-    :mask-close="false"
-    :draggable="false"
-    :width="730"
-    :transfer="false"
-    @after-leave="handleClosed">
-    <div class="title" slot="tools">
-      <span>{{$t('筛选条件')}}</span>
-      <bk-input class="filter-input" v-model.trim="filter" clearable :placeholder="$t('请输入关键字搜索')"></bk-input>
-    </div>
-    <section class="property-selector">
-      <div class="group"
-        v-for="group in renderGroups"
-        :key="group.id">
-        <h2 class="group-title">
-          {{group.name}}
-          <span class="group-count">（{{group.children.length}}）</span>
-        </h2>
-        <ul class="property-list clearfix">
-          <li class="property-item fl"
-            v-for="property in group.children"
-            :key="property.bk_property_id">
-            <bk-checkbox class="property-checkbox"
-              :disabled="disabledProperties[group.id].includes(property.bk_property_id)"
-              :checked="isChecked(property)"
-              @change="handleToggleProperty(property, ...arguments)">
-              <span v-bk-tooltips.top-start="{
-                disabled: !disabledProperties[group.id].includes(property.bk_property_id),
-                content: $t('该字段不支持搜索')
-              }">
-                {{property.bk_property_name}}
-              </span>
-            </bk-checkbox>
-          </li>
-        </ul>
-      </div>
-      <cmdb-data-empty v-if="renderGroups.length === 0" class="empty-content" slot="empty"
-        :stuff="dataEmpty"
-        @clear="handleClearFilter"></cmdb-data-empty>
-    </section>
-    <footer class="footer" slot="footer">
-      <i18n class="selected-count"
-        v-if="selected.length"
-        path="已选择条数"
-        tag="div">
-        <template #count><span class="count">{{selected.length}}</span></template>
-      </i18n>
-      <div class="selected-options">
-        <bk-button theme="primary" @click="confirm">{{$t('确定')}}</bk-button>
-        <bk-button theme="default" @click="close">{{$t('取消')}}</bk-button>
-      </div>
-    </footer>
-  </bk-dialog>
+  <addCondition
+    ref="addConditionComp"
+    :selected="selected"
+    :disabled-property-map="disabledProperties"
+    :models="models"
+    :property-map="propertyMap">
+  </addCondition>
 </template>
 
 <script>
   import { mapGetters } from 'vuex'
   import FilterStore from './store'
-  import throttle from 'lodash.throttle'
   import { PROPERTY_TYPES } from '@/dictionary/property-constants'
+  import addCondition from '@/components/add-condition'
+
   export default {
+    components: {
+      addCondition
+    },
     data() {
       return {
-        filter: '',
-        isShow: false,
-        selected: [...FilterStore.selected],
-        throttleFilter: throttle(this.handleFilter, 500, { leading: false }),
-        renderGroups: [],
-        dataEmpty: {
-          type: 'empty',
-          payload: {
-            defaultText: this.$t('暂无数据')
-          }
-        }
+        selected: [...FilterStore.selected]
       }
     },
     computed: {
@@ -154,6 +101,13 @@
         })
           .sort((groupA, groupB) => sequence.indexOf(groupA.id) - sequence.indexOf(groupB.id))
       },
+      models() {
+        return this.groups.map(group => ({
+          id: group.id,
+          bk_obj_name: group.name,
+          bk_obj_id: group.id
+        }))
+      },
       disabledProperties() {
         const disabledPropertyMap = {}
         this.groups.forEach((group) => {
@@ -164,149 +118,19 @@
         return disabledPropertyMap
       }
     },
-    watch: {
-      filter: {
-        immediate: true,
-        handler(value) {
-          this.dataEmpty.type = value ? 'search' : 'empty'
-          this.throttleFilter()
-        }
-      }
-    },
     methods: {
-      handleFilter() {
-        if (!this.filter.length) {
-          this.renderGroups = this.groups
-        } else {
-          const filteredGroups = []
-          const filter = this.filter.toLowerCase()
-          this.groups.forEach((group) => {
-            const properties = group.children.filter((property) => {
-              const name = property.bk_property_name.toLowerCase()
-              return name.indexOf(filter) > -1
-            })
-            if (properties.length) {
-              filteredGroups.push({
-                ...group,
-                children: properties
-              })
-            }
-          })
-          this.renderGroups = filteredGroups
-        }
-      },
-      isChecked(property) {
-        return this.selected.some(target => target.id === property.id)
-      },
-      handleToggleProperty(property, checked) {
-        if (checked) {
-          this.selected.push(property)
-        } else {
-          const index = this.selected.findIndex(target => target.id === property.id)
-          index > -1 && this.selected.splice(index, 1)
-        }
-      },
       async confirm() {
-        FilterStore.updateSelected(this.selected)
-        FilterStore.updateUserBehavior(this.selected)
+        const selected = this.$refs?.addConditionComp?.localSelected ?? this.selected
+        FilterStore.updateSelected(selected)
+        FilterStore.updateUserBehavior(selected)
         this.close()
+      },
+      close() {
+        this.$emit('closed')
       },
       handleClosed() {
         this.$emit('closed')
       },
-      open() {
-        this.isShow = true
-      },
-      close() {
-        this.isShow = false
-      },
-      handleClearFilter() {
-        this.filter = ''
-      }
     }
   }
 </script>
-
-<style lang="scss" scoped>
-    .title {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        vertical-align: middle;
-        line-height: 31px;
-        font-size: 24px;
-        color: #444;
-        padding: 15px 0 0 24px;
-        .filter-input {
-            width: 240px;
-            margin-right: 45px;
-        }
-    }
-    .property-selector {
-        margin: 0 -24px -24px 0;
-        height: 350px;
-        @include scrollbar-y;
-        position: relative;
-        .empty-content{
-            position: absolute;
-            top:50%;
-            left:50%;
-            transform: translate(-50%,-50%);
-        }
-    }
-    .group {
-        margin-top: 15px;
-        .group-title {
-            position: relative;
-            padding: 0 0 0 15px;
-            line-height: 20px;
-            font-size: 15px;
-            font-weight: bold;
-            color: #63656E;
-            &:before {
-                content: "";
-                position: absolute;
-                left: 0;
-                top: 3px;
-                width: 4px;
-                height: 14px;
-                background-color: #C4C6CC;
-            }
-            .group-count {
-                color: #C4C6CC;
-                font-weight: normal;
-            }
-        }
-    }
-    .property-list {
-        padding: 10px 0 6px 0;
-        .property-item {
-            width: 33%;
-        }
-    }
-    .property-checkbox {
-        display: block;
-        margin: 8px 20px 8px 0;
-        @include ellipsis;
-        /deep/ {
-            .bk-checkbox-text {
-                max-width: calc(100% - 25px);
-                @include ellipsis;
-            }
-        }
-    }
-    .footer {
-        display: flex;
-        .selected-count {
-            font-size: 14px;
-            line-height: 32px;
-            .count {
-                color: #2DCB56;
-                padding: 0 4px;
-            }
-        }
-        .selected-options {
-            margin-left: auto;
-        }
-    }
-</style>
