+++ src/ui/src/components/ui/permission/embed-permission.vue
@@ -17,7 +17,7 @@
         <div>
           <i18n path="抱歉您没有查看权限">
             <template #link>
-              <bk-button class="text-btn"
+              <bk-button class="text-btn no-permission"
                 text
                 theme="primary"
                 @click="handleApplyPermission">
@@ -52,6 +52,11 @@
         .text-btn {
             font-size: 14px;
             height: auto;
+        }
+        .no-permission {
+            display: block;
+            width: 100%;
+            margin-top: 10px;
         }
     }
 </style>
