+++ src/ui/src/i18n/lang/cn.json
@@ -1114,7 +1114,7 @@
     "已选择N个模板：": "已选择 {count} 个模板：",
     "仅能在业务模块下新增": "仅能在业务模块下新增",
     "创建实例": "创建实例",
-    "抱歉您没有查看权限": "抱歉您没有查看权限，{link}",
+    "抱歉您没有查看权限": "抱歉您没有查看权限{link}",
     "您还未XXX": "您还未{action}{resource}，{link}",
     "搜索内容为空": "搜索内容为空",
     "主机自动应用": "主机自动应用",
